refactor: Use store cache for pod datastore
