+++ central/pod/datastore/datastore.go
@@ -2,13 +2,10 @@ package datastore
 
 import (
 	"context"
-	"testing"
 
 	"github.com/stackrox/rox/central/pod/datastore/internal/search"
-	"github.com/stackrox/rox/central/pod/store/cache"
-	pgStore "github.com/stackrox/rox/central/pod/store/postgres"
+	pgStore "github.com/stackrox/rox/central/pod/datastore/internal/store/postgres"
 	piDS "github.com/stackrox/rox/central/processindicator/datastore"
-	piFilter "github.com/stackrox/rox/central/processindicator/filter"
 	plopDS "github.com/stackrox/rox/central/processlisteningonport/datastore"
 	v1 "github.com/stackrox/rox/generated/api/v1"
 	"github.com/stackrox/rox/generated/storage"
@@ -34,21 +31,7 @@ type DataStore interface {
 
 // NewPostgresDB creates a pod datastore based on Postgres
 func NewPostgresDB(db postgres.DB, indicators piDS.DataStore, plops plopDS.DataStore, processFilter filter.Filter) (DataStore, error) {
-	store, err := cache.NewCachedStore(pgStore.New(db))
-	if err != nil {
-		return nil, err
-	}
+	store := pgStore.New(db)
 	searcher := search.New(store, pgStore.NewIndexer(db))
 	return newDatastoreImpl(store, searcher, indicators, plops, processFilter), nil
-}
-
-// GetTestPostgresDataStore provides a datastore connected to postgres for testing purposes.
-func GetTestPostgresDataStore(t testing.TB, pool postgres.DB) (DataStore, error) {
-	processIndicatorStore, err := piDS.GetTestPostgresDataStore(t, pool)
-	if err != nil {
-		return nil, err
-	}
-	plopStore := plopDS.GetTestPostgresDataStore(t, pool)
-	processIndicatorFilter := piFilter.Singleton()
-	return NewPostgresDB(pool, processIndicatorStore, plopStore, processIndicatorFilter)
 }
