+++ central/pod/datastore/internal/store/postgres/gen.go
@@ -1,0 +1,3 @@
+package postgres
+
+//go:generate pg-table-bindings-wrapper --type=storage.Pod --cached-store --search-category PODS --references storage.Deployment
