--- central/pod/store/postgres/gen.go
@@ -1,3 +1,0 @@
-package postgres
-
-//go:generate pg-table-bindings-wrapper --type=storage.Pod --search-category PODS --references storage.Deployment
